Move docker task definitions into a plugin
