+++ deployment/build.gradle.kts
@@ -1,8 +1,10 @@
-import java.io.IOException
+import com.hom.DockerBuildAndSaveTask
+import com.hom.DockerTask
 
 plugins {
     base
     id("com.palantir.git-version") version "3.0.0"
+    id("hom.docker")
 }
 
 val frontendBuild by configurations.creating {
@@ -43,70 +45,9 @@ tasks {
         outputTar.set(layout.buildDirectory.file("$name/${dockerImageName.replace(":", "--")}.tar"))
     }
 
-    val dockerImport by registering(DockerTask::class) {
-        args.add("import")
-        args.add(dockerBuildAndSave.flatMap { task -> task.outputTar.map { it.asFile.absolutePath } })
-        args.add(dockerBuildAndSave.flatMap { it.imageName })
-    }
-
     val dockerRun by registering(DockerTask::class) {
         dependsOn(dockerBuildAndSave)
         args.set(listOf("run", "-p", "8080:8080", "--name", "hit-or-miss"))
         args.add(dockerBuildAndSave.flatMap { it.imageName })
-    }
-}
-
-@DisableCachingByDefault(because = "Not all Docker tasks have a clear output")
-abstract class DockerTask @Inject constructor(private val execOps: ExecOperations) : DefaultTask() {
-
-    @get:Input
-    abstract val args: ListProperty<String>
-
-    @get:[Optional Input]
-    abstract val workingDir: Property<File>
-
-    @TaskAction
-    fun runDockerCommand() {
-        val dockerCommandArgs = listOf("docker") + args.get()
-        logger.info("Running '${dockerCommandArgs.joinToString(" ")}'.")
-        val dockerWorkingDir = workingDir
-        execOps.exec {
-            if (dockerWorkingDir.isPresent) {
-                workingDir(dockerWorkingDir.get())
-            }
-            commandLine(dockerCommandArgs)
-        }
-    }
-}
-
-@CacheableTask
-abstract class DockerBuildAndSaveTask @Inject constructor(private val execOps: ExecOperations) : DefaultTask() {
-
-    @get:Input
-    abstract val buildDir: Property<File>
-
-    @get:Input
-    abstract val imageName: Property<String>
-
-    @get:OutputFile
-    abstract val outputTar: RegularFileProperty
-
-    @TaskAction
-    fun runDockerCommands() {
-        // Build
-        val dockerBuildCommand = listOf("docker", "build", "-t", imageName.get(), ".")
-        execOps.exec {
-            workingDir(buildDir.get())
-            commandLine(dockerBuildCommand)
-        }
-        // Save
-        val outputTarFile = outputTar.get().asFile
-        if (!outputTarFile.parentFile.exists() && !outputTarFile.parentFile.mkdirs()) {
-            throw IOException("Couldn't create output directory: '${outputTarFile.parentFile.absolutePath}'.")
-        }
-        val dockerSaveCommand = listOf("docker", "save", imageName.get(), "--output", outputTarFile.absolutePath)
-        execOps.exec {
-            commandLine(dockerSaveCommand)
-        }
     }
 }
