+++ gradle/plugins/settings.gradle.kts
@@ -1,1 +1,3 @@
 rootProject.name = "plugins"
+
+include(":docker")
